Do not use glibc-internal header sys/cdefs.h (#210)
The `sys/cdefs.h` is an internal header of `glibc` and is not meant to
be used by user code.  The macros needed from it can be trivially
implemented where required, relying on this header being present
excludes the OpenThread Border Router from being used on systems where
`glibc` is not the used C library (e.g. OpenWRT, where `uclibc` or
`musl` is standard).

https://wiki.musl-libc.org/faq.html#Q:-When-compiling-something-against-musl,-I-get-error-messages-about-%3Ccode%3Esys/cdefs.h%3C/code%3E