disable nginx instead of apache
The BeagleBone Debian distribution has changed from Jessie to Strecth.
One of the changes in that migration was moving the server for the
Cloud9 IDE to nginx from apache.

One of the scripts called by the setup disables running webservers to
allow the OpenThread Border Router to run it's own server. This updates
the disable script to point to the current webserver on the BeagleBone.