[web] topology visualization (#519)
This is the first PR of OpenThread Network Manager.

In general, it implements a new module -- OTBR-REST, which is directly
embedded in OTBR-AGENT and provides RESTful API for a new feature of
WEB UI -- Topology Visualization.

Topology Visualization aims to collect diagnostic information of all
nodes in current network, visualize their relations and provide the
display of all the details.

The frontend part is implemented based on the frontend of OTBR-WEB for
the time being. We will then implement all the functionality OTBR-WEB
provides and transplant the frontend static file of OTBR-WEB in this
module.