[WIP] add HIDL server support
This CL implements the HIDL server on the Android platform. The HIDL
server provides IThread and IThreadSettings HIDL interfaces for the
HIDL client.
The IThread interface is implemented by the HIDL server to provides
all control operations of the Thread network.
The IThreadSettings interface is mainly called by the HIDL server to
send sensitive settings to Thread System service. The Thread System
servie stores the sensitive settings to a securely area.