[mdns] add vendor mdnssd API support
In Android, the otbr-agent runs in vendor partition. The mDNSResponder
runs in the system partition. Hence, the otbr-agent can't directlly call
the mDNSResponder API. It has to call the mDNSResponder APIs through AIDL
interface.

This CL adds vendor mdnssd APIs to allow the user to implement the
functions to access the mDNSResponder APIs in Android repo.