[mdns] Fix pointer to object on stack in TxtList
TxtList is a list of DNS TXT records to publish, represented
as memory ranges for keys and values of the records.
While memory ranges for the values point directly to a
otSrpServerService object's member whose lifetime is longer
than lifetime of the TxtList, memory ranges for the keys
pointed to a member of otDnsTxtEntryIterator structure,
allocated on stack during the translation.

It resulted in publishing garbage TXT records.

Fix the issue by using std::string to store the keys instead
of raw pointers/lengths.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>