[mbedtls] enable MBEDTLS_SSL_KEEP_PEER_CERTIFICATE define
The mbedTLS 2.17 introduced a new option called
`MBEDTLS_SSL_KEEP_PEER_CERTIFICATE` which allows saving RAM by not
keeping the peer's certificate after completion of the handshake. This
define is enabled by default in reference configuration but not in the
`mbedtls-config.h`. This functionality is needed when building the
OpenThread Posix application (more specifically
`Dtls::GetPeerCertificateBase64` method).