[android] fix android build check for mDNSResponder (#860)
Currently in the github action for checking android build, the env
OTBR_MDNS doesn't take effect because the script is actually run in a
docker image. Thus when building it, libmdnsd and PublisherMDnsSd are
not compiled. This commit adds OTBR_MDNS env variable to the docker
and FLAGS, INCLUDES for using mDNSResponder in ot-br-posix.