[adproxy] re-register all services when the mDNS publisher becomes ready (#888)
When mDNS daemon restarts, the mDNS library may need to re-advertise
all hosts and services. This commit extracts this functionality and
makes it a public method.

Also, on kReady, SRP server will publish all registered hosts and
services. This is helpful when the advertising proxy reconnects to the
mDNS provider.