[border-agent] name meshcop service after "VendorName_ProductName"  (#946)
The recommended format of service instance is VendorName
ProductName._meshcop._udp_.local.. However, it's quite tricky to
handle the space in service name properly. Also, the space in the name
will be converted to \032 after the registration, which is not a very
elegant form. Therefore, set
VendorName_ProductName._meshcop._udp_.local. as the default service
name.