[d-bus] implement D-Bus API UpdateVendorMeshCopTxtEntries (#1081)
This commit implements a D-Bus API for updating the entries in TXT
record of MeshCoP service.

The API caller provides a list of (key, value) pairs, which contains
all the entries it wants to update. Then otbr-agent will re-publish
the MeshCoP service with the updated TXT record. If the list is empty,
it will reset the TXT record, discarding changes made by previous API
calls.