[mdns-avahi] fix potential memory access issue (#1131)
We are passing a pointer to the `ServiceSubscription` to Avahi methods
like `avahi_service_browser_new` and `avahi_service_resolver_new` as
the userdata.  However, since `ServiceSubscription` objects were
directly stored in a `vector<ServiceSubscription>`, any update to the
vector causes the `ServiceSubscription` objects to move, leading to
potential invalid pointer access.

This commit fixes this issue by storing `ServiceSubscription` pointers
in the vector.

This commit also fixes the bug that the `ServiceSubscription` and
`HostSubscription` objects may be freed in resolved callbacks.