[mdns] add a callback for each service and host registraion (#1113)
This commit refactors the mDNS publisher to receive and bind a unique
callback for each service/host registration. The result of the
service/host registration is solely delivered back to the caller via
the callback.

This commit also changes to not auto-rename the service or host name
when conflicts are detected but returns a `OTBR_ERROR_DUPLICATED`
error and let the caller to resolve name conflicts on its own.