[border-agent] don't treat `OTBR_ERROR_ABORTED` as an error in logs (#1307)
otbr-agent generates `WARNING` level logs when there are frequent
meshcop service updates. Such a log line is generated because an
ongoing meshcop service registration is canceled for some reason (a
newer registration preempts the previous registration, or border agent
wants to stop advertising the service). However, this behavior is
working as expected and shouldn't be regarded as an error. Hence I
lowered the log level to `INFO` in such cases.