[dns64] set DNS64 default value to 0 (#1291)
This commit sets DNS64 default value to 0 for all platforms. The DNS64
flag currently installs and configures bind9. When it's ON, the DNS-SD
server will not work properly.

It also adds `python3` as a Docker dependency since some tests require
it. It seems that python3 was silently installed along with bind9.