[platform] de-init application in `otPlatReset` and verify `Publisher` state in public methods (#1339)
When `otPlatReset` is called on `otbr-agent`, it should de-init the
old `otbr::Application` so that the resources will be properly
disposed, otherwise there will be dangling resources. Also,
`otbr::Application` should de-init the TREL module when itself is
de-initing.

One example is that if we call `ot-ctl factoryreset`, the old
`otbr::Application` won't be destructed so that the publisher will
keep the old mDNS service registrations until the process exits. In
such a situation, there will be two different MeshCoP services
published by the same `otbr-agent` process.

Also, this commit adds a verification at `Publisher` public methods to
ensure that `Publisher` is at `kReady` state.