[meshcop] append last 2 bytes of ext addr to service instance name (#1370)
This commit appends last 2 bytes of extended address to the MeshCoP
service instance name. This is used for reducing service instance name
conflicts when there are two BRs of the same model connected to the
same wifi.

An example of the new service instance name is `OpenThread
BorderRouter #B76D`.