[nat64] separate Tayga configuration from NAT64 configuration (#1435)
This commit separates Tayga configuration from NAT64 configuration.

It extracts the Tayga commands into helper functions.

It adds a new env `NAT64_SERVICE`. The current default value is
`tayga` and can be updated later (e.g. `openthread` for the native
service).

If `NAT64_SERVICE`=`tayga`, Tayga is installed and triggers the
relevant commands. Otherwise it's not.

If `NAT64_SERVICE`!=`tayga`, the native NAT64 service is expected to
be used. An `iptables MASQUERADE` is configured for the ipv4 address
pool of the native NAT64 service. It can be updated later if needed.