[dbus] provide Feature Flag update API for DBus (#1563)
Feature Flag definition and its update DBus API provides the mechanism
to define and maintain feature flags for Thread over time. Use
OTBR_FEATURE_FLAGS to turn on/off this change. It is designed to
support:
* The ground truth of feature flag definition and their suggested
  default values in case of feature rollback.
* API and flag change compatibility between OTBR layer and upper
  layer.

The flags are defined in proto definition. The proto runtime is more
lightweight than DBus' serialization so that we believe it is safe to
include it in the library.