[openthread] disable power calibration (#1702)
The commit https://github.com/openthread/ot-br-posix/pull/1701 and
https://github.com/openthread/ot-br-posix/pull/1696 depends on each
other. The `Depends-on` doesn't work on ot-br-posix, it causes both
commits `1701` and `1696` can't pass the github action check.

This commit temporarily disable power calibration to let the commit
`1696` to pass the github action check. After the commit `1696` is
merged, update the commit `1701` to enable the power calibration again.