[logging] detailed logging over feature experiment (#1678)
This commit enables detailed logging over feature experiment.

Background: On some test devices the default log level is NOTE, which
may not be sufficient for debugging when a bug is produced. So we
would like to be able to set a detailed Thread log level in an
automated fashion, which will be applied at the beginning of the
thread process.

A feature flag 'enable_detailed_logging‘ of bool type is supported to
enable default detailed logging level 'INFO'. An additional field
'detailed_logging_level‘ is used to specify the log level.