[cmake] sort the options in `openthread/CMakeLists.txt` (#1759)
This commit re-arranges the options in `openthread/CMakeLists.txt` to
keep them sorted and re-use same `if` block. It does not add or
remove any option. The goal is to make it easier to read and see
which OT features(CMake options) are enabled in `ot-br-posix` build.

This commit also changes variable types from `BOOL` to `STRING` to
match the definition in OT core.