[mdns] keep avahi service resolver until instance is removed (#1783)
There is a difference between avahi and dns-sd browse operation. An
update in a discovered instance doesn't trigger any avahi browse
events, while for dns-sd rmv/add events are triggered.

In the current avahi mdns design, after service browse we resolve the
discovered instance once, when instance is resolved the resolver is
removed.

Trel uses dnssd browse to discover peers, when a discovered peer
changes ipv6 address, trel would not know the change and still try to
send packets to the old address.