[rest] return border agent ID (#1864)
* [rest] return node information when detached

Ignore error return of otThreadGetLeaderData(). It only returns an error
when the Thread border router is detached, which we don't want to
consider as an error condition. Instead, make sure the NodeInfo struct
is cleared. This makes sure that we don't leak stack data.

* [rest] return border agent ID

Add BaId to the node JSON object and add a dedicated endpoint
/node/ba-id to retrive the border agent ID.