[mdns] let avahi use `RecordBrowser` to resolve addresses of a service (#1875)
Avahi only returns one of the host's addresses in the service
resolution callback. However, the address may be link-local so it may
not be preferred from Thread's perspective. We want to go through the
complete list of addresses associated with the host and choose a
routable address. Therefore we use `RecordBrowser` which can do this
job.