[rest] initialize dataset (#1892)
When there is already a dataset, and setting properties using the PUT
method and an incomplete JSON dataset, some fields of the dataset
variable might not be initialized. This leads to uninitialized data
being used as active dataset.

Make sure we initialize the dataset variable. This makes sure that all
components of the dataset are considered not set.