[utils] update SuccessOrExit for error tracking
This commit updates the `SuccessOrExit` macro to accept an action
before goto "exit". It allows us to add a log when error happens.
For example, we can write:

```
SuccessOrExit((error = writeDatasetToSettings()), otbrLogErr("Failed to
write dataset to settings"));
```

So that we know exactly which step a method fails at.

------

This commit also prints the error/status code for macro `SuccessOrDie`.