[mdns] update `TxtEntry`  to handle boolean attribute (#1992)
This commit updates `TxtEntry`, `EncodeTxtData()`, and `DecodeTxtData()`
to support boolean attributes, which are encoded as `key` string
without the `=` character. It also ensures that `EncodeTxtData()`
will generate valid TXT data (containing a single zero byte) when
given an empty `TxtList`. Additionally, this commit updates
`AdvertisingProxy::MakeTxtList()` to use `DecodeTxtData()`, and adds
a test to validate the encoder and decoder behavior.