[mdns-mdnssd] handle deallocated `ServiceRef`s as processing run loop
This commit updates `PublisherMDnsSd::Process()`, which processes
ready `DNSServiceRef`s from the main run loop. We now track the list
of ready service refs in a member variable, `mServiceRefsToProcess`.

As we iterate over the list of `mServiceRefsToProcess`, the call to
`DNSServiceProcessResult()` can itself invoke callbacks into
`PublisherMDnsSd` and OT, which in turn, may change the state of
Publisher and potentially trigger a previously valid `ServiceRef` in
the list to be deallocated.

This commit adds `HandleServiceRefDeallocating()`, which is called
whenever a `ServiceRef` is being deallocated. From this method, we
update the entry in the `mServiceRefsToProcess` list to `nullptr` to
avoid calling `DNSServiceProcessResult()` on an already freed
`ServiceRef`.