[thread] refine `RetrieveTelemetryData` to populate metrics with best effort (#2076)
Best effort means, for a given telemetry, if its retrieval has error,
it is left unpopulated and the process continues to retrieve the
remaining telemetries instead of the immediately return. The error
code OT_ERRROR_FAILED will be returned if there is one or more
error(s) happened in the process.

This change helps to retrieve telemetries with best effort when
needed.