[controller] add ncp host (#2329)
This commit adds the `NcpHost` class which is the implementation of
`ThreadController` under NCP case.

This commit updates `ThreadController::Create` to create `NcpHost`
under the NCP case, allowing `otbr-agent` to start when the
co-processor is a NCP, though no functionalities have been implemented
yet.

This commit also adds a CI workflow to use expect script to test
otbr-agent in NCP case. For this PR, a simple test is added to start
otbr-agent with NCP (dry run) and check the co-processor version.