[github-actions] update the build directory for otbr-agent (#2366)
This commit updates the build directory of otbr-agent in ncp_mode CI
tests to a different path. This allows us to build a docker image from
the same repo later. Currently if we build otbr-agent in default path
and then we build a docker image, it will find existing cmake cache
and report an error.

Building docker image is to test border routing features of NCP in
simulation environment.