[host] add network properties class (#2387)
This commit adds a new interface `NetworkProperties` to provide a
unified, synchronized way to access network properties for both NCP
and RCP.
* For RCP, the property getters are implemented by OT APIs with an OT
  instance.
* For NCP, the property getters are implemented by the replica of the
  data that comes from NCP.

This is important for adding NCP support because in some cases the
network properties are required by some functionality and we don't
want to make an asynchronous call simply to get the property. As one
example, in AdvertisingProxy, we need to the MeshLocal Prefix to check
if an address is mesh local address. Currently this is done by OT API,
which is not available under NCP mode. However, the information is
contained in the active dataset and NCP will always notify the host
about the change. With `NetworkProperties`, we can always get these
properties in a synchronized way.

This commit also replaces a few usage of `otThreadGetDeviceRole` by
`NetworkProperties`. In this way we will gradually remove the direct
usage of OT APIs and mInstance in ot-br-posix modules. In this way we
can develop for both NCP&RCP more easily.