[tests] enhance the test script of ncp mode to speed up local developing (#2394)
This commit enhances the `ncp_mode` bash test script to enhance local
developing.

This commit adds argument parsing for two arguments:
1. --build-ot-sim This allows to avoid building ot-cli-ftd and
   ot-ncp-ftd every time running the test.
2. --one-test This allows to specify one specific test to run.

This commit enhances local developing and doesn't change the behavior
in CI.

This commit also separates the building of ot-cli-ftd and ot-ncp-ftd
for future consideration. In the future, we will enable things like
`OT_BORDER_ROUTING` for ot-ncp-ftd but we won't enable it for
reference ot-cli-ftd.