[posix] add posix netif module (#2410)
This commit adds a posix platform module for NCP mode. Under NCP mode,
the platform operation won't depend OT posix platform implementation
and will use the posix platform implementation added in this PR
instead.

This commit only adds the netif module and support creating the Thread
network interface (wpan0) and setting the MTU. The other
functionalities will be added in following PRs. The netif module
hasn't been integrated with the NcpHost. In this PR, the posix
platform is added as a library and has its own unit test (also
implemented by gtest).

The netif implementations are different on different platforms: linux,
apple, bsd... This commit only implements it on linux and adds an empty
implementation for other posix platforms.