[ncp] update the frame handling of ncp spinel (#2416)
This commit updates the general logic of the handling of incoming
spinel frames in NcpSpinel.

Currently the dispatching depends on the incoming key (what we
actually received), which brings some implementation troubles. For
example, the response of a get command and a set command could
sometimes be the same thing: the property.

In this commit, we use `mCmdTable` together with `mWaitingKeyTable` to
record each sent command and do the handling based on `mCmdTable[tid]`
and `mWaitingKeyTable[tid]` (what we expect to receive). This makes it
much easier to further add more handlings later.