[dbus] add dbus Migrate API for NCP (#2427)
This commit adds the 'Migrate' Dbus API under interface.

This commit only adds the dbus interface with a pseudo
implementation. (not calling `ThreadHost` API to do the actual
migration) This will be added later.

The 'Migrate' API shares some common implementation as
`ThreadHelper::AttachAllNodesTo` method. So the PR does a minimum
refactor to expose the process to validate/append PENDING_TIMESTAMP
and DELAY tlvs from `ThreadHelper` so that `DBusThreadObjectNcp` can
reuse the code.

This commit adds a small expect test to check the handling of input
dataset.