[posix] add unicast address update in netif module (#2431)
This commit implements `UpdateIp6UnicastAddresses` method for
`Posix::Netif` module to add/remove unicast address on the wpan
interface. A unit test is added together.

The method `UpdateIp6UnicastAddresses` takes a vector of Ip6 addresses
as input (because the address update from the NCP is the entire
address table) and update the entire address list to the
interface. However, it doesn't do it by removing all and then adding
all. Instead, it compares the existing unicast addresses against the
new addresses to decide which addresses to add/remove and then issue
netlink requests to do the changes.