[posix] add netif TUN Ip6 Sending (#2452)
In specific, when a packet from the system is sent to the wpan
interface, the Netif module will do the Ip6 sending. In this commit
the actual sending (let NCP send the IP6 packet) hasn't been
implemented and an empty implementation is used.

The commit also adds a unit test to verify IP packet sent through wpan
interface can be successfully handled in netif.