[netif] refactor Netif and NcpSpinel (#2479)
`Netif::Ip6Send` depends on the implementation of `NcpSpinel`. And
there are more methods to come. (For example,
`SubcribeMulticastAddress`). Currently we pass the Ip6Send
implementation to `Netif` in `Netif::Init` through a std::function
object. But then, each time we add a new method, we need to update the
signature of `Netif::Init`(making it very long) and the calling of
`Netif::Init`, which is not preferred. This commit adds a dependencies
class `NetifDependencies` to encapsulate all the implementation
functions for Netif and let `NcpSpinel` implement this interface.