[netif] add wpan multicast address subscribe/unsubscribe for NCP (#2478)
This commit allows the applications to join/leave multicast groups in
the Thread network. The functionality is implemented by listening to
MLD reports sent by kernel. And then the Netif module will send the
join/leave request to the NCP and let NCP join/leave the multicast
group.

In this commit, `NcpSpinel::Ip6MulAddrUpdateSubscription` is
implemented as empty. This will be implemented in a following
commit. This commit adds a unit test case to verify that the MLD
reports can be successfully received and handled.