[infra if] add platform infra if module (#2501)
In this commit, InfraIf module only implements SetInfraIf method. This
method is intended to be called by NcpHost during initialization or
through ThreadHost API (for example, for the android case, where the
infra if is ready later). This method will initialize the infra if and
the border routing module on NCP.