[infra if] implement infra if state update (#2533)
This commit implements Infra If state update in the InfraIf module.

This commit enables the `InfraIf` module to listen to the state change
of the infrastructure interface and then notify the change to the
NCP. This is achieved by listening to netlink events. The
implementation is basically the same as OT posix InfraIf.

This commit adds a unit test case to verify
`InfraIfDependency::SetInfraIf` is called correctly when the InfraIf
state changes.