[infra if] implement infra if send icmp6 nd (#2534)
This commit implements `SendIcmp6Nd` in the `InfraIf` module.

The method is used to send an ICMPv6 ND message (on host) when NCP
calls `otPlatInfraIfSendIcmp6Nd`. In this commit, the method hasn't
been called. It will be later integrated with `NcpSpinel`. The
implementation of this method is the same as OT posix implementation.