[infra if] add infra if icmp6 nd receiving (#2547)
This commit implements Icmp6Nd receiving in the InfraIf module.

The method is used to receive ICMPv6 ND messages (on host) and forward
them to the NCP. The implementation of this method is the same as OT
posix implementation. In this commit, the forwarding to NCP hasn't
been implemented. It will be later integrated with NcpSpinel.

This commit also adds a test case to verify that `HandleIcmp6Nd` is
called correctly when the infrastructure interface receives an ICMP6
ND message.