[tests] add unit test for RcpHost SetCountryCode (#2564)
This commit adds a unit test for `RcpHost::SetCountryCode`.

The commit also fixes an issue of current usage of `FakePlatform`. In
each gtest case, we should get a new OT Instance. Otherwise the OT
instance used in last gtest case will have some remaining tasks which
are invalid to process.