[thread-host] add common network properties (#2598)
This commit adds a few common network properties to `ThreadHost`.

The direct usage of these properties is in Android platform when
refreshing the state. With this commit, that process can work for both
NCP & RCP. This commit only implements the Get methods for RCP. The
NCP version will be implemented later.