[ncp] integrate SRP Server and Advertising Proxy for NCP (#2624)
This commit integrates the mDNS Publisher with NcpHost to make
advertising proxy work under NCP mode.

The commit adds an expect test to verify the advertising proxy
function under NCP mode. To do the test, `avahi-browser` is used. So
it's added into the bootstrap of the CI for ncp_mode.

As an abstraction: 
1. The commit adds `NcpSpinel::SrpServerSetEnabled` and
   `NcpSpinel::SrpServerSetAutoEnableMode` to control the SRP server
   on the NCP.
2. The commit adds the handling of `SPINEL_PROP_DNSSD_HOST`,
   `SPINEL_PROP_DNSSD_SERVICE` and `SPINEL_PROP_DNSSD_KEY_RECORD` to
   help NCP to publish the dnssd entries (by using the mDNSPublisher
   on the host).