[build] allow using shared boost libraries
Do not unconditionally set `Boost_USE_STATIC_LIBS` but instead default
it to `true`.

This allows building with shared Boost libraries, which can be useful
for OS distribution packaging, for example.

The default is the same if not defined, so this is a backwards
compatible change: unless explicitly set to `false`, static Boost
libs will still be used.