[pd] disable PD daemon which caused mDNS disruptions (#2662)
This commit addresses the mDNS disruptions caused by the
dhcp6_pd_daemon restarting dhcpcd on reference device. As a
workaround, the daemon is disabled, and PD is enabled by default in
the dhcpcd.conf file. This ensures the reference device as BR, always
requests a prefix delegation upon network interface configuration.