[build] allow disabling systemd service unit install (#2652)
Add `INSTALL_SYSTEMD_UNIT` which can explicitly be set to `false`
to disable installing the service unit even if systemd is available.

If not defined, it defaults to true, i.e. the systemed service unit
is created. As a result, this change is backwards compatible: it's
required to _opt-out_ of systemd unit installation.