[mdns-publisher] use `FlagsMoreComing` in `Mdns::Publisher`
This commit enhances `Mdns::Publisher` using mDNSResponder to utilize
the `kDNSServiceFlagsMoreComing` flag. When this flag is present in a
callback, it indicates that additional results are forthcoming. In
such cases, we delay reporting the result to collect additional
addresses.

This improves host and service resolution by preventing
early "resolved" callbacks after receiving only the first IPv6
address. This allows the publisher to gather the additional
addresses (from cache or all the addresses included in a query
responses).