[test] fix socat version to 1.7.4.4 (#2706)
This commit changes the installation of socat from using 'apt' to
manual installation.

We found the latest socat version 1.8.0.3 has an issue when running
with OTBR docker. To avoid the error caused by version and facilitate
local development, the commit provides a script to install socat of
version `1.7.4.4` and replaces the installation in CI and scripts..

BTW, the installation in `install_openthread_binraries` is unnecessary
because `install_common_dependencies` will be done anyway.