[border-agent] add `Init()` and `Deinit()` for `BorderAgent`  (#2745)
This commit implements `Init()` and `Deinit()` for `BorderAgent`. This
is to support the case on Android that we need to properly
re-initialize the `BorderAgent` after leaving a network. Previously
the Thread state callback was registered in the constructor so it goes
away after `Deinit` of `ThreadHost` and will never get back.