[border-agent] add UDP Proxy module for Thread commissioning (#2717)
This commit adds a UDP Proxy module to do Thread commissioning under
NCP architecture.

This commit is one of sub-task to support Border Agent under NCP
mode. The UDP Proxy will also be used to do Thread commisssioning
under RCP architecture and replace `OT_PLATFORM_UDP`.

Based on the OT border agent state changes
(HandleBorderAgentStateChange), the module simply starts/stops
listening on an UDP ephemeral port and if starts, sets a mapping
between OT border agent UDP port and the port on host. And then it
does bidrectional UDP forwarding:
* `Dependencies::UdpForward` will be implemented as OT UDP forwarding
  to OT core.
* `SendToPeer` will send a UDP packet to the remote peer using the
  local host port (that maps to OT border agent port) as the source
  port.

The module hasn't been integrated with the application and put in
use. This commit adds a unit test for the module.

This module can also be used for other cases.