[border-agent] refactor epskc state change handling (#2755)
This commit refactors the Epskc state change handling in BorderAgent.

The commit adds an API `AddEphemeralKeyStateChangedCallback` in
`ThreadHost` to add ephemeral key state callback so that the
application rely on this API and unbind `BorderAgent` from OT
instance. The intention is to make BorderAgent able to work in NCP
mode as well.