[test] fix flaky unit test of test_netif (#2761)
This commit fixes a flaky unit test in test netif
(`WpanIfRecvIp6PacketCorrectly_AfterReceivingFromNetif`).

The cause is that if we call `recvfrom` right after
`Netif::Ip6Receive`, it is possible that the system hasn't passed the
message and the program hangs on `recvfrom`. Then the program will
hang forever.

I've seen the failure in CI for many times. I've verified the test
locally for 10 times with this change.