[dbus] lower the log level in `ReplyOtResult` on error to reduce false alarms  (#2778)
The current behavior of `ReplyOtResult` is that it will print an
`ERROR` logging when the error is not `OT_ERROR_NONE`. This may not be
appropriate because the error may not indicate a fault in the
system. For example, it shouldn't be concerning if
`GetActiveDatasetTlvs` returns `NotFound` while the TBR hasn't got any
credentials.

This commit lowers the log level to INFO to reduce false alarms.