[dbus] refactor dbus module to unbind it with border agent feature (#2775)
This commit refactors the DBusAgent module to unbind it with the
BORDER_AGENT_ENABLE flag as well as simply the code.

Currently the `mDBusAgent` is guarded with `#if
OTBR_ENABLE_DBUS_SERVER &&
OTBR_ENABLE_BORDER_AGENT`. `OTBR_ENABLE_BORDER_AGENT` shouldn't be a
precondition to enable the DBus server.

We have the current code because the DBusAgent relies on the
BorderAgent when `OTBR_ENABLE_BORDER_AGENT` is enabled. However the
code to inject the `BorderAgent` to DBusAgent is cumbersome: We have
to write two versions of constructor or init for many classes. So as
of now we only have one constructor which always accepts a
`BorderAgent`.

This commit introduces a helper class `DBus::DependentComponents`. The
constructor and init method will always accept this class. And we
control the fields in this class by using flags like
`OTBR_ENABLE_BORDER_AGENT`. Thus we don't need to write two versions
of constructors and can enable DBusAgent without having to enable
BorderAgent.