[ncp] add meshcop commissioning in NCP mode (#2785)
This commit implements meshcop commissioning in NCP mode.

This commit also introduces ot-commissioner installation in the
`ncp_mode` script and extends the expect test of ncp_border_agent to
verify the commissioning function in NCP mode.

This commit adds `SetUdpForwardToHostCallback`to `ThreadHost` and let
`ThreadHost` implements `UdpProxy::Dependencies`. As of now `RcpHost`
has an empty implementation. In later commits we will implement it for
`RcpHost` and repalce the platform UDP.